Issue: Major refactoring needed for frontend Solution:   * Change all react component file extensions to .jsx   * Change app.css to app.module.css   * Remove unnecessary files auto-generated by create-react-app   * Change all directory names to start with lower case & remove white space in the name   * Change logo file type from png to favicon.ico   * Enforce one directory per component   * Move image directories to src/assets/image Risk: React app do not run (further refactoring required) Reviewed By: Alex, Aiden, Jason
