+++ frontend/src/components/stickInsectsContainer/stickInsectContainer.jsx
